+++ pytsite/contact_form/__init__.py
@@ -9,12 +9,20 @@ __license__ = 'MIT'
 
 
 def _init():
-    from pytsite import assetman, lang, tpl, http_api
+    from pytsite import assetman, lang, tpl, http_api, permission, settings
+    from . import _settings_form
 
+    # Resources
     assetman.register_package(__name__)
     lang.register_package(__name__)
     tpl.register_package(__name__)
+
+    # HTTP API endpoints
     http_api.register_package('contact_form', 'pytsite.contact_form.http_api')
 
+    # Settings form
+    permission.define_permission('contact_form.settings.manage', 'pytsite.contact_form@manage_contact_form', 'app')
+    settings.define('contact_form', _settings_form.Form, 'pytsite.contact_form@contact_form', 'fa fa-paper-plane',
+                    'contact_form.settings.manage')
 
 _init()