+++ pytsite/widget/_button.py
@@ -51,7 +51,7 @@ class Button(_base.Abstract):
 
         self._html_em.content = self.get_val()
         if self._icon and not self._html_em.children:
-            self._html_em.append(_html.I(cls='fa fa-fw fa-' + self._icon))
+            self._html_em.append(_html.I(cls=self._icon))
 
         for k, v in self._data.items():
             self._html_em.set_attr('data_' + k, v)